refactor: Extract RDF data models to ConstructResponseRdfData
Move type aliases and data classes (RdfResources, Statements, PredicateObjects, FlatStatements, RdfPropertyValues, RdfData, ValueRdfData, ResourceWithValueRdfData, MappingAndXSLTransformation, MainResourcesAndValueRdfData) from ConstructResponseUtilV2 to a new dedicated object ConstructResponseRdfData. Update imports across MainQueryResultProcessor and SearchResponderV2.
